% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{allow_shiny}
\alias{allow_shiny}
\title{Simple check to determine if code is being run in RStudio with the shiny runtime}
\usage{
allow_shiny()
}
\description{
Simple check to determine if code is being run in RStudio with the shiny runtime
}
