% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ames_sampling_dist.R
\name{ames_sampling_dist}
\alias{ames_sampling_dist}
\title{Run the interactive ames sampling distribution shiny app to 
illustrate sampling distributions using variables from the \code{\link{ames
}}}
\usage{
ames_sampling_dist()
}
\description{
Run the interactive ames sampling distribution shiny app to 
illustrate sampling distributions using variables from the \code{\link{ames
}}
}
\examples{
\dontrun{ames_sampling_dist}
}
