% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arbuthnot_lab.R
\docType{data}
\name{arbuthnot}
\alias{arbuthnot}
\title{FIXME - Arbuthnot's Birth Records}
\format{A data frame with 82 rows and 3 variables:
\describe{
  \item{year}{}
  \item{boys}{}
  \item{girls}{}
}}
\source{
\url{}
}
\usage{
arbuthnot
}
\description{
Blah blah blah
}
\keyword{datasets}

