% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atheism.R
\docType{data}
\name{atheism}
\alias{atheism}
\title{Atheism in the world data}
\format{
A tbl_df with 88032 rows and 3 variables:
\describe{
  \item{nationality}{Country of the individual surveyed.}
  \item{response}{A categorical variable with two levels: atheist and non-atheist.}
  \item{year}{Year in which the person was surveyed.}
  }
}
\source{
Global Index of Religiosity and Atheism. WIN-Gallup International Press. 2012.
}
\usage{
atheism
}
\description{
Survey results on atheism across several countries and years. Each row
represents a single respondent.
}
\keyword{datasets}
