% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_sim.R
\name{bandit_sim}
\alias{bandit_sim}
\title{Run the Bandit Simulation shiny app}
\usage{
bandit_sim()
}
\description{
Simulate data from a two armed-bandit (two slot machines) by clicking
on the images for Machine 1 or Machine 2 and guess/learn which machine 
has the higher probability of winning as the number of 
outcomes of wins and losses accumulate.
}
