% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_streak.R
\name{calc_streak}
\alias{calc_streak}
\title{Calculate hit streaks.}
\usage{
calc_streak(x)
}
\arguments{
\item{x}{A data frame or character vector of hits (\code{"H"}) and misses (\code{"M"}).}
}
\value{
A data frame with one column, \code{length}, containing the length of each hit streak.
}
\description{
Calculate hit streaks.
}
\examples{
data(kobe_basket)
calc_streak(kobe_basket$shot)
barplot(table(calc_streak(kobe_basket$shot)), col = "blue", 
        xlab = "Streak length", ylab = "Frequency")
test1 <- c("H","M","M","H","H","M","M","M","H","M")
test2 <- c("H","M","M","H","H","M","M","M","H","H")
test3 <- c("M","M","M","H","H","M","M","M","H","H")
calc_streak(test1)
barplot(table(calc_streak(test1)), col = "green",
        xlab = "Streak length", ylab = "Frequency")
calc_streak(test2)
calc_streak(test3)

}
