% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_one_mean_sim.R
\name{ci_single_mean_sim}
\alias{ci_single_mean_sim}
\title{Confidence interval for a single mean, using simulation}
\usage{
ci_single_mean_sim(y, conf_level, y_name, boot_method, nsim, seed,
  show_var_types, show_summ_stats, show_res, show_eda_plot, show_inf_plot)
}
\arguments{
\item{y}{Numerical vector}

\item{conf_level}{Confidence level desired (between 0 and 1)}

\item{y_name}{Name of the y variable, to be used in output and plot axes}

\item{boot_method}{bootstrap method; "perc" (percentile) or "se" (standard error)}

\item{nsim}{number of simulations}

\item{seed}{seed to be set, default is NULL}

\item{show_res}{print results, set to verbose by default}

\item{show_eda_plot}{print EDA plot, set to verbose by default}

\item{show_inf_plot}{print inference plot, set to verbose by default}

\item{show_var_type}{print variable types, set to verbose by default}

\item{show_sum_stats}{print summary stats, set to verbose by default}
}
\description{
Confidence interval for a single mean, using simulation
}

