% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{is_shiny_runtime}
\alias{is_shiny_runtime}
\title{Simple check to determine if code is being run in RStudio with the shiny runtime}
\usage{
is_shiny_runtime()
}
\description{
Simple check to determine if code is being run in RStudio with the shiny runtime
}
