% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot_hand_lab.R
\docType{data}
\name{kobe_basket}
\alias{kobe_basket}
\title{FIXME - Kobe Bryant's shots made}
\format{A data frame with 133 rows and 6 variables:
\describe{
  \item{vs}{}
  \item{game}{}
  \item{quarter}{}
  \item{time}{}
  \item{description}{}
  \item{shot}{}
}}
\source{
\url{}
}
\usage{
kobe_basket
}
\description{
Blah blah blah
}
\keyword{datasets}

