% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arbuthnot_lab.R
\docType{data}
\name{present}
\alias{present}
\title{FIXME - Present Birth Records}
\format{A data frame with 74 rows and 3 variables:
\describe{
  \item{year}{}
  \item{boys}{}
  \item{girls}{}
}}
\source{
\url{}
}
\usage{
present
}
\description{
Blah blah blah
}
\keyword{datasets}

