% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wage.R
\docType{data}
\name{wage}
\alias{wage}
\title{Wage data}
\format{A tbl_df with with 935 rows and 17 variables:
\describe{
  \item{wage}{weekly earnings (dollars)}
  \item{hours}{average hours worked per week}
  \item{iq}{IQ score}
  \item{kww}{Knowledge of world work score}
  \item{educ}{years of education}
  \item{exper}{years of work experience}
  \item{tenure}{years with current employer}
  \item{age}{age in years}
  \item{married}{=1 if married}
  \item{black}{=1 if black}
  \item{south}{=1 if live in south}
  \item{urban}{=1 if live in a Standard Metropolitan Statistical Area }
  \item{sibs}{number of siblings}
  \item{brthord}{birth order}
  \item{meduc}{mother's education (years)}
  \item{feduc}{father's education (years)}
  \item{lwage}{natural log of wage}
}}
\source{
Jeffrey M. Wooldridge (2000). Introductory Econometrics: A Modern Approach. South-Western College Publishing.
}
\usage{
wage
}
\description{
The data were gathered as part of a random sample of 935 respondents throughout the United States.
}
\keyword{datasets}

