% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ames_sampling_dist.R
\name{ames_sampling_dist}
\alias{ames_sampling_dist}
\title{Simulate Sampling Distribution}
\usage{
ames_sampling_dist()
}
\description{
Run the interactive ames sampling distribution shiny app to 
illustrate sampling distributions using variables from the `ames`
dataset.
}
\examples{
if (interactive()) { 
  ames_sampling_dist()
}
<<<<<<< HEAD

=======
>>>>>>> 7bb9e0a254cbbd7f3d25aac9a6849c9c5acf73f1
}
