% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present.R
\docType{data}
\name{present}
\alias{present}
\title{Male and female births in the US}
\format{A tbl_df with 74 rows and 3 variables:
\describe{
  \item{year}{year, ranging from 1940 to 2013}
  \item{boys}{number of male births}
  \item{girls}{number of female births}
}}
\source{
Data up to 2002 appear in Mathews TJ, and Hamilton BE. 2005. Trend
analysis of the sex ratio at birth in the United States. National Vital
Statistics Reports 53(20):1-17. Data for 2003 - 2013 have been collected
from annual National Vital Statistics Reports published by the US Department of 
Health and Human Services, Centers for Disease Control and Prevention, 
National Center for Health Statistics.
}
\usage{
present
}
\description{
Counts of the total number of male and female births in the United States from
1940 to 2013.
}
\keyword{datasets}

