% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsr.R
\docType{package}
\name{statsr}
\alias{statsr}
\alias{statsr-package}
\title{statsr: A companion package for the Coursera Statistics with R specialization}
\description{
See https://github.com/StatsWithR/statsr for more information.
}

